feat: add abstract zap contract
