+++ .gitignore
@@ -8,3 +8,4 @@ reports/
 
 # Node/npm
 node_modules/
+.DS_Store
